Minor adjustment to code. Added Documentation for Visual Classes and World Class

+++ Visuals/BonusView.h
@@ -7,13 +7,16 @@
 
 #include "EntityView.h"
 
+/// Class that models the visual representation of a Bonus.
+/// \param framecounter : int : a counter to slow down the visual updates of the Bonus' frame texture. Only relevant to the JumpPad Bonus.
 class BonusView : public EntityView {
-    float frameCounter;
+    int frameCounter;
 public:
     BonusView();
     explicit BonusView(const sf::Texture& texture);
     BonusView(const sf::Texture& texture, const sf::IntRect& rect);
 
+    /// Updates the sprite's visual frame if frameCounter reaches 0.
     void update();
 
 };
